refactor: move menu enums constants -> constants/menus
