+++ src/main/java/christmas/constants/menus/Dessert.java
@@ -1,14 +1,13 @@
-package christmas.constants;
+package christmas.constants.menus;
 
-public enum Drink implements Menu {
-    ZERO_COLA("제로콜라", 3000),
-    RED_WINE("레드와인", 60000),
-    CHAMPAGNE("샴페인", 25000);
+public enum Dessert implements Menu {
+    CHOCOLATE_CAKE("초코케이크", 15000),
+    ICE_CREAM("아이스크림", 5000);
 
     private final String name;
     private final int price;
 
-    Drink(String name, int price) {
+    Dessert(String name, int price) {
         this.name = name;
         this.price = price;
     }
